refactor(gui): share form view boilerplate via BaseFormView
