+++ src/gui/flight/view.py
@@ -1,61 +1,7 @@
-from PySide6.QtWidgets import (
-    QFormLayout,
-    QGroupBox,
-    QHBoxLayout,
-    QLineEdit,
-    QPushButton,
-    QVBoxLayout,
-    QWidget,
-)
-
+from gui.common.base_form_view import BaseFormView
 from gui.flight.types import FLIGHT_TEXT_FIELDS
 
 
-class FlightFormView(QWidget):
+class FlightFormView(BaseFormView):
     def __init__(self) -> None:
-        super().__init__()
-        self.form_inputs: dict[str, QLineEdit] = {}
-
-        layout = QVBoxLayout(self)
-        layout.setContentsMargins(0, 0, 0, 0)
-        layout.addWidget(self._build_form_group())
-        layout.addLayout(self._build_crud_row())
-
-    def _build_form_group(self) -> QGroupBox:
-        form = QFormLayout()
-        for field in FLIGHT_TEXT_FIELDS:
-            line_edit = QLineEdit()
-            if field == "Date":
-                line_edit.setPlaceholderText("YYYY-MM-DDTHH:MM:SS")
-            self.form_inputs[field] = line_edit
-            form.addRow(f"{field}:", line_edit)
-
-        group = QGroupBox("Flight Details")
-        group.setLayout(form)
-        return group
-
-    def _build_crud_row(self) -> QHBoxLayout:
-        self.create_btn = QPushButton("Create")
-        self.update_btn = QPushButton("Update")
-        self.delete_btn = QPushButton("Delete")
-        self.clear_btn = QPushButton("Clear")
-
-        row = QHBoxLayout()
-        for button in (
-            self.create_btn,
-            self.update_btn,
-            self.delete_btn,
-            self.clear_btn,
-        ):
-            row.addWidget(button)
-        row.addStretch()
-        return row
-
-    def read_payload(self) -> dict[str, str]:
-        return {
-            field: widget.text().strip() for field, widget in self.form_inputs.items()
-        }
-
-    def clear(self) -> None:
-        for widget in self.form_inputs.values():
-            widget.clear()
+        super().__init__(text_fields=FLIGHT_TEXT_FIELDS, group_name="Flight Details")
